+++ public/index.html
@@ -24,12 +24,12 @@
       work correctly both with client-side routing and a non-root public URL.
       Learn how to configure a non-root public URL by running `npm run build`.
     -->
-    <title>React App</title>
+    <title>Instagram Clone</title>
   </head>
   <body style="background-color: #fafafa;">
     <noscript>You need to enable JavaScript to run this app.</noscript>
     <div id="header"></div>
-    <div id="root"></div>
+    
     <!--
       This HTML file is a template.
       If you open it directly in the browser, you will see an empty page.
