+++ src/AppHeader/app-header-style.css
@@ -29,6 +29,7 @@
 
 .SearchBox{
   position: relative;
+  
 }
 .SearchBoxLogo{
   position: absolute;
@@ -42,10 +43,17 @@
   background-color: #fafafa;
   outline: none;
   border:1px solid #eceaea;
-  border-radius: 2px;
+  border-radius: 5px;
+  padding:2px;
   padding-left: 25px;
   font-weight: 300;
   font-size: 14px;
+  width: 250px;
+}
+.profilepicture{
+  border-radius: 50%;
+  min-width: 30px;
+  max-height:30px;
 }
 
 
@@ -65,6 +73,9 @@
     display: none;
   }
   .loginButton{
+    display: none;
+  }
+  .app-header-profile-area{
     display: none;
   }
 }
