+++ src/index.css
@@ -1,9 +1,3 @@
-*{
-  object-fit: contain;
-  margin: 0%;
-  text-align: left;
-}
-
 body{
   background-color: #fafafa;
 }
